Force a sane DNS config for virt cluster
The default DNS servers in the Ubuntu Vagrant box is causing occasional
name resolution issues, and trusting the base image to provide sane DNS
is a little problematic anyway.

This change adds a step to force a known-good DNS server to be
configured as part of the virtual cluster turnup process.

This is kinda a hack but it appears to work, having tested for both
Ubuntu and CentOS.